requirements.txt
amqp==5.0.9
appdirs==1.4.4
asgiref==3.5.2
attrs==22.1.0
audioread==2.1.9
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
bz2file==0.98
CacheControl==0.12.10
cachetools==4.2.4
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
charset-normalizer==2.0.7
ciseau==1.0.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.11.0
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.1
django-admin-interface==0.20.0
django-admin-sortable2==2.1.2
django-allauth==0.45.0
django-colorfield==0.7.2
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
django-storages==1.12.3
django-tinymce==3.4.0
djangochannelsrestframework==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
epub==0.5.2
epub-conversion==1.0.15
et-xmlfile==1.1.0
fcm-django==1.0.11
filelock==3.3.2
firebase-admin==5.2.0
fonttools==4.33.3
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
huggingface-hub==0.1.2
hyperlink==21.0.0
HyperPyYAML==1.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.3
kombu==5.2.3
librosa==0.9.2
llvmlite==0.38.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.2
msgpack==1.0.3
noisereduce==2.0.1
numba==0.55.2
numpy==1.21.4
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.2
pandas==1.4.3
Pillow==8.4.0
pooch==1.6.0
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.3.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.7.2
sentencepiece==0.1.96
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
SoundFile==0.10.3.post1
speechbrain==0.5.10
sqlparse==0.4.2
threadpoolctl==3.1.0
torch
torchaudio
tqdm==4.62.3
Twisted==22.4.0
txaio==22.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zope.interface==5.4.0
